Feature/cytoscape graph

+++ NoteAssistant.KnowledgeGraph.Backend/Services/IngestionStore.cs
@@ -5,16 +5,16 @@ namespace NoteAssistant.KnowledgeGraph.Backend.Services;
 
 public sealed class IngestionStore
 {
-    private readonly ConcurrentDictionary<int, IngestionStatusDto> _statuses = new();
-    private readonly ConcurrentDictionary<int, GraphIngestionPlan> _plans = new();
-    private readonly ConcurrentDictionary<int, IngestionExecutionLogDto> _executionLogs = new();
+    private readonly ConcurrentDictionary<long, IngestionStatusDto> _statuses = new();
+    private readonly ConcurrentDictionary<long, GraphIngestionPlan> _plans = new();
+    private readonly ConcurrentDictionary<long, IngestionExecutionLogDto> _executionLogs = new();
 
     public void Upsert(IngestionStatusDto status)
     {
         _statuses[status.DocumentId] = status with { UpdatedAt = DateTimeOffset.UtcNow };
     }
 
-    public IngestionStatusDto? Get(int documentId)
+    public IngestionStatusDto? Get(long documentId)
     {
         _statuses.TryGetValue(documentId, out var status);
         return status;
@@ -25,18 +25,18 @@ public sealed class IngestionStore
         _plans[plan.DocumentId] = plan;
     }
 
-    public GraphIngestionPlan? GetPlan(int documentId)
+    public GraphIngestionPlan? GetPlan(long documentId)
     {
         _plans.TryGetValue(documentId, out var plan);
         return plan;
     }
 
-    public void SaveExecutionLog(int documentId, IngestionExecutionLogDto log)
+    public void SaveExecutionLog(long documentId, IngestionExecutionLogDto log)
     {
         _executionLogs[documentId] = log;
     }
 
-    public IngestionExecutionLogDto? GetExecutionLog(int documentId)
+    public IngestionExecutionLogDto? GetExecutionLog(long documentId)
     {
         _executionLogs.TryGetValue(documentId, out var log);
         return log;
